Move method builder to Emit folder
